requirements.txt
selenium==4.8.0
pytest==7.2.1
allure-pytest==2.9.45
pytest-xdist
requests~=2.31.0
jsonschema~=4.17.3